requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.40
PyJWT==2.8.0
requests==2.32.0
# --- testing ---
pytest==8.2.1        # test runner
